Change delta time type into examples.
Former-commit-id: 6af42fc06d78fb634be9fcbbff5f54de18396013

+++ example/multi-surfaces/src/Main.cpp
@@ -136,7 +136,7 @@ int main(int argc, char** argv)
 		camera->component<PerspectiveCamera>()->aspectRatio((float)w / (float)h);
 	});
 
-	auto enterFrame = canvas->enterFrame()->connect([&](Canvas::Ptr canvas, uint time, uint deltaTime)
+	auto enterFrame = canvas->enterFrame()->connect([&](Canvas::Ptr canvas, uint time, float deltaTime)
 	{
 		//spotLight->component<Transform>()->matrix()->appendRotationY(0.01f);
 		sceneManager->nextFrame();
